Separate interface for Swagger annotations
This will ensure that the controller is more focused and that all Swagger documentation is stored in one place.
