--- crown-court-proceeding/src/main/java/uk/gov/justice/laa/crime/crowncourt/controller/CrownCourtProceedingController.java
@@ -1,89 +1,0 @@
-package uk.gov.justice.laa.crime.crowncourt.controller;
-
-import io.swagger.v3.oas.annotations.Operation;
-import io.swagger.v3.oas.annotations.Parameter;
-import io.swagger.v3.oas.annotations.media.Content;
-import io.swagger.v3.oas.annotations.media.Schema;
-import io.swagger.v3.oas.annotations.responses.ApiResponse;
-import io.swagger.v3.oas.annotations.tags.Tag;
-import jakarta.validation.Valid;
-import lombok.RequiredArgsConstructor;
-import lombok.extern.slf4j.Slf4j;
-import org.springframework.http.MediaType;
-import org.springframework.http.ResponseEntity;
-import org.springframework.web.bind.annotation.*;
-import uk.gov.justice.laa.crime.annotation.DefaultHTTPErrorResponse;
-import uk.gov.justice.laa.crime.crowncourt.builder.CrownCourtDTOBuilder;
-import uk.gov.justice.laa.crime.crowncourt.dto.CrownCourtDTO;
-import uk.gov.justice.laa.crime.crowncourt.model.request.ApiProcessRepOrderRequest;
-import uk.gov.justice.laa.crime.crowncourt.model.request.ApiUpdateApplicationRequest;
-import uk.gov.justice.laa.crime.crowncourt.model.response.ApiProcessRepOrderResponse;
-import uk.gov.justice.laa.crime.crowncourt.model.response.ApiUpdateApplicationResponse;
-import uk.gov.justice.laa.crime.crowncourt.model.response.ApiUpdateCrownCourtOutcomeResponse;
-import uk.gov.justice.laa.crime.crowncourt.service.ProceedingService;
-import uk.gov.justice.laa.crime.crowncourt.validation.CrownCourtDetailsValidator;
-
-@Slf4j
-@RestController
-@RequiredArgsConstructor
-@RequestMapping("api/internal/v1/proceedings")
-@Tag(name = "Crown Court Proceeding", description = "Rest API for Crown Court Proceeding.")
-public class CrownCourtProceedingController {
-
-    private final ProceedingService proceedingService;
-
-    private final CrownCourtDetailsValidator crownCourtDetailsValidator;
-
-    private CrownCourtDTO preProcessRequest(ApiProcessRepOrderRequest request) {
-        return CrownCourtDTOBuilder.build(request);
-    }
-
-    @PostMapping(produces = MediaType.APPLICATION_JSON_VALUE)
-    @Operation(description = "Process Rep Order Data")
-    @ApiResponse(responseCode = "200",
-            content = @Content(mediaType = MediaType.APPLICATION_JSON_VALUE,
-                    schema = @Schema(implementation = ApiProcessRepOrderRequest.class)
-            )
-    )
-    @DefaultHTTPErrorResponse
-    public ResponseEntity<ApiProcessRepOrderResponse> processRepOrder(
-            @Parameter(description = "Process Crown Rep Order",
-                    content = @Content(mediaType = MediaType.APPLICATION_JSON_VALUE,
-                            schema = @Schema(implementation = ApiProcessRepOrderRequest.class)
-                    )
-            ) @Valid @RequestBody ApiProcessRepOrderRequest request) {
-
-        CrownCourtDTO requestDTO = preProcessRequest(request);
-        return ResponseEntity.ok(
-                proceedingService.processRepOrder(requestDTO)
-        );
-    }
-
-    @PutMapping(produces = MediaType.APPLICATION_JSON_VALUE)
-    @Operation(description = "Update Crown Court Application")
-    @ApiResponse(responseCode = "200",
-            content = @Content(mediaType = MediaType.APPLICATION_JSON_VALUE,
-                    schema = @Schema(implementation = ApiUpdateApplicationRequest.class)
-            )
-    )
-    @DefaultHTTPErrorResponse
-    public ResponseEntity<ApiUpdateApplicationResponse> updateApplication(@Valid @RequestBody ApiUpdateApplicationRequest request) {
-        CrownCourtDTO crownCourtDTO = preProcessRequest(request);
-        return ResponseEntity.ok(proceedingService.updateApplication(crownCourtDTO));
-    }
-
-
-    @PutMapping(value = "/update-crown-court" , produces = MediaType.APPLICATION_JSON_VALUE)
-    @Operation(description = "Update Crown Court")
-    @ApiResponse(responseCode = "200",
-            content = @Content(mediaType = MediaType.APPLICATION_JSON_VALUE,
-                    schema = @Schema(implementation = ApiUpdateApplicationRequest.class)
-            )
-    )
-    @DefaultHTTPErrorResponse
-    public ResponseEntity<ApiUpdateCrownCourtOutcomeResponse> update(@Valid @RequestBody ApiUpdateApplicationRequest request) {
-        CrownCourtDTO crownCourtDTO = preProcessRequest(request);
-        crownCourtDetailsValidator.checkCCDetails(crownCourtDTO);
-        return ResponseEntity.ok(proceedingService.update(crownCourtDTO));
-    }
-}
